Move inclusion proof code to new file and adjust tests
