+++ src/lib.rs
@@ -14,3 +14,6 @@ pub mod utils;
 
 #[cfg(test)]
 mod tests;
+
+#[cfg(test)]
+mod testing_utils;
